Cast (int) in printf for 64-bit magma_int_t
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@3286 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ src/dsyevdx_2stage.cpp
@@ -273,7 +273,7 @@ magma_dsyevdx_2stage(char jobz, char range, char uplo,
     if(lda2>n){
         #ifndef NO_WARNING    
         printf("--------------------------------------------------------------\n");
-        printf("  warning matrix too small N=%d NB=%d, calling lapack on CPU  \n",n,nb);
+        printf("  warning matrix too small N=%d NB=%d, calling lapack on CPU  \n", (int) n, (int ) nb);
         printf("--------------------------------------------------------------\n");
         #endif
         lapackf77_dsyevd(&jobz, &uplo, &n,
